diagrams/scripts/requirements.txt
click~=7.1
pyyaml~=5.3
scikit-image~=0.18
